# blockchain: Switch to FindSpentTicketsInBlock.