--- a/blockchain/chainio.go
+++ b/blockchain/chainio.go
@@ -1277,7 +1277,7 @@ func (b *BlockChain) createChainState() error {
 	// Create a new node from the genesis block and set it as the best node.
 	genesisBlock := dcrutil.NewBlock(b.chainParams.GenesisBlock)
 	header := &genesisBlock.MsgBlock().Header
-	node := newBlockNode(header, nil, nil, nil)
+	node := newBlockNode(header, &stake.SpentTicketsInBlock{})
 	node.inMainChain = true
 	b.bestNode = node
 
@@ -1439,10 +1439,8 @@ func (b *BlockChain) initChainState() error {
 
 		// Create a new node and set it as the best node.  The preceding
 		// nodes will be loaded on demand as needed.
-		blk := dcrutil.NewBlock(&block)
 		header := &block.Header
-		node := newBlockNode(header, ticketsSpentInBlock(blk),
-			ticketsRevokedInBlock(blk), voteBitsInBlock(blk))
+		node := newBlockNode(header, stake.FindSpentTicketsInBlock(&block))
 		node.inMainChain = true
 		node.workSum = state.workSum