--- a/blockchain/internal_test.go
+++ b/blockchain/internal_test.go
@@ -18,7 +18,6 @@ import (
 	"time"
 
 	"github.com/decred/dcrd/blockchain/stake"
-	"github.com/decred/dcrd/chaincfg/chainhash"
 	"github.com/decred/dcrd/wire"
 )