--- a/blockchain/stakeversion_test.go
+++ b/blockchain/stakeversion_test.go
@@ -22,7 +22,7 @@ import (
 func newFakeChain(params *chaincfg.Params) *BlockChain {
 	// Create a genesis block node and block index index populated with it
 	// for use when creating the fake chain below.
-	node := newBlockNode(&params.GenesisBlock.Header, nil, nil, nil)
+	node := newBlockNode(&params.GenesisBlock.Header, &stake.SpentTicketsInBlock{})
 	node.inMainChain = true
 	index := make(map[chainhash.Hash]*blockNode)
 	index[node.hash] = node
@@ -53,7 +53,7 @@ func newFakeNode(parent *blockNode, blockVersion int32, stakeVersion uint32, bit
 		Timestamp:    timestamp,
 		StakeVersion: stakeVersion,
 	}
-	node := newBlockNode(header, nil, nil, nil)
+	node := newBlockNode(header, &stake.SpentTicketsInBlock{})
 	node.parent = parent
 	node.workSum.Add(parent.workSum, node.workSum)
 	return node