First try at improving #913.

This adds a new function `FindSpentTicketsInBlock` that extracts all spent and revoked tickets, plus the vote bits of a given block and uses that function instead of the individual functions when loading a new node into the blockchain.

This improves startup time a bit, as the `isSSGen`, `isSSRtx`, and `determineTxType` are somewhat slow, as they need to decode the output script of the transactions. Looping over the transactions once and doing a single test is faster.

The first commit adds a couple of instrumentation lines (to be removed before final merge).

Numbers for my computer:
```
| variation     | mainnet | testnet |
| with patch    |   23.7s |   42.6s |
| without patch |   35.9s |  118.0s |
```